Fix formatting of mercurial revisions with Python 3.8
In Python 3.8, attempting to use formatting on a bytestring with
dictionary parameters will end up looking for the keys as bytestrings
rather than unicode. This results in a KeyError. This change fixes it so
we do the format entirely as `str` and then encode to `bytes`.

Testing Done:
- Verified that the encoding worked correctly.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/11135/