Do not rewrite hg base_commit_id if not necessary
If the Mercurial base_commit_id does not contain any relative
information we do not need to rewrite it. This saves an expensive call
of the hg binary.

Testing Done:
Used a base_commit_id with relative character and saw the rewrite. Used
a normal base_commit_id and saw that it did not rewrite.

Reviewed at https://reviews.reviewboard.org/r/11145/