Fix `rbt version` tests to use the same python interpreter as the tests.
Instead of just invoking `rbt version` to test the new version output,
we need to make sure that it uses the same Python version as
runtests.py (which may or may not be the same as what `python` points to
inside the current virtualenv).

Trivial fix.

Testing Done:
Ran unit tests on Python 2.7 and 3.8