Add support for a reviewboard.repository_name counter in Perforce.
Historically, we've provided support for specifying a Review Board
server URL as a Perforce counter key/value, useful for very large
deployments that don't want to juggle `.reviewboardrc` files. This has
still left RBTools to scan for repositories, comparing paths, which is
slow.

This change introduces some base level support for `SCMClient`s to
provide a repository name as part of the `RepositoryInfo`, which will be
used as a fallback if an explicit repository name is not specified
during repository and review request lookup.

For the Perforce side, a new `reviewboard.repository_name` counter is
now supported, which provides the value for this new field.

While we still recommend a `.reviewboardrc` for deployments, this should
help RBTools work faster and more reliably with very large, old,
established Perforce repositories.

Testing Done:
Unit tests pass.

Set the counter in a test repository and verified Review Board was trying
to use it for posting changes and displaying status.

Reviewed at https://reviews.reviewboard.org/r/11030/