Fix `rbt status` returning blank summary for draft review requests
When displaying the summary for the review request, the code attempts to
get the summary from `request.summary`, which is null for draft reviews.
In this change, we added a condition on where to get the summary - if
the draft exists, we get the draft summary; otherwise we get it from the
published review request.

Testing Done:
1. Reproduced issue in 4855, confirmed the summary for a draft review
shows as expected after the code change:
```
+--------+-----------------------------------------------------+---------------+
| Status |                   Review Request                    |    Branch     |
+========+=====================================================+===============+
| Draft  | r/11174 - fix 4855:rbtools 1.0.2 returns blank      | release-1.0.x |
|        | summary for draft review requests                   |               |
+--------+-----------------------------------------------------+---------------+
```
2. Ran the unit tests using `./tests/runtests.py`, all passed.

Bugs closed: 4855

Reviewed at https://reviews.reviewboard.org/r/11174/