Fix issues testing rbt with the right environment.
The unit tests for the `rbt` command weren't always running the correct
version of RBTools. The main test suite was running whatever `rbt` was
in the path with whatever version of Python it was installed with, which
didn't guarantee correct results, while the `--version` tests ran with
the current Python but also ran the first build found in the path.

These could produce wrong results, and in fact have in my own test
environment. To fix this, we now run the bundled copy specifically, and
we now always do so with the current version of Python. This ensures a
reliable test run, and also allows us to consolidate the two test suites
for this command.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/11275/