Fix rbt land --squash when landing a review request.
`rbt land --squash` was correctly squashing commits when landing a local
branch, but not when loading a review request with multiple commits. All
the work to handle squashing a patch was already done, but we just
weren't plumbing the `--squash` argument to `rbt patch` to trigger it.

This is a simple change that adds the missing argument, fixing squashing
of multi-commit review requests.

Testing Done:
Successfully landed a multi-commit review request with `--squash`.

Reviewed at https://reviews.reviewboard.org/r/11296/