Modernize the RBTools for Windows installer for Python 3 and 64-bit.
This updates the installer to install Python 3.8 instead of 2.7. This
was chosen instead of the newly-released Python 3.9 due to its
stability, though we may move to 3.9 in a future release.

The installer also now installs a 64-bit build of Python (and of our
wrappers) instead of 32-bit. This should be fine for most users. Some
work was done to aid in creating a 32-bit installer later, if we get
user feedback requesting one.

There's also some improvements to the Wix installer files. Strings are
localized, comments were added, and some sections are cleaned up.

Testing Done:
Built the installer and tested it on Windows. Successfully installed
RBTools, and verified that I could run it from a new DOS shell.

Verified the files were placed in `C:\Program Files` and not
`C:\Program Files (x86)`.

Reviewed at https://reviews.reviewboard.org/r/11299/