Build 32-bit and 64-bit Windows installers.
We've historically shipped 32-bit Windows installers for RBTools, which
include a 32-bit Python. With RBTools 2.0, we're shipping a 64-bit
installer, but it's possible there are users who still need 32-bit.

The previous change for the installer set up some of the Wix rules to
allow for 32-bit and 64-bit builds. This completes that work by updating
our `build-installer.bat` to pull down both architectures for Python and
then to build two installers: `RBTools-2.0-32bit.exe` and
`RBTools-2.0-64bit.exe`.

This mostly involved performing some of the same work multiple times in
a batch file. There's also a couple fixes for issues hit when trying to
run the build process, and the removal of a probably unnecessary
condition check in the installer that prevents 32-bit installers from
being run on 64-bit (which work in practice and shouldn't be blocked).

Testing Done:
Built the installers. Tested both on Windows, ensuring they installed
correctly, contained the expected files, and that RBTools could be run.

Reviewed at https://reviews.reviewboard.org/r/11300/