Add documentation on all of the .reviewboardrc configuration options.
One of the big things the RBTools docs have lacked was a complete
breakdown of all available `.reviewboardrc` options. We pretty much left
all but the main ones as an exercise to the reader. The best we offered
was a reference to an option name in the docs for command line options.

This change updates our user and repository docs to list all options
that are available, with type information, examples, and relevant
commands.

I've taken a best guess at where each option should go. Some are likely
to be pretty user-specific, and so those have been placed in the user
docs. The rest live in the repository docs.

Each option makes use of a new ReST crossref type called `rbtconfig`.
This can be used to reference a particular config option no matter where
it is in the docs. The generated list of option docs now use this to
link to the appropriate config docs.

Testing Done:
Built the docs. Checked for broken errors and links.

Reviewed at https://reviews.reviewboard.org/r/11303/