Fix up the testing information and related projects in the README.
The README was a bit out-of-date, and wasn't updated prior to 2.0. It
contained some old information on setting up and running unit tests,
links to our GitHub repositories for related projects instead of our
project pages, and old project names.

This change modernizes those docs, pointing to the current instructions
for setting up tests and ensuring our links are valid and named
correctly.

Testing Done:
Opened the docs in a Markdown viewer. Checked the formatting and the links.

Reviewed at https://reviews.reviewboard.org/r/11310/