Change "vobstag" to "vobtag" in ClearCase implementation.
The ClearCase client was entirely third-party contributed, and now that
I'm more familiar with how ClearCase works, one of the things that's odd
about it is the use of the name `vobstag`. I suspect this was meant to
mean the "VOB's tag", but it's still quite odd and doesn't really match
the way we name variables. This change updates it to use `vobtag`
instead.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/11498/