Add support for VersionVault
We will soon be adding a new SCMTool implementation to Power Pack to connect to
HCL VersionVault and IBM ClearCase, supplanting the old community-driven
ClearCase support. This change adds RBTools-side support for this.

This is based on code delivered to us in partnership with HCL, but has had
significant changes. The largest change is that instead of having a completely
separate SCMClient implementation in RBTools, this is built into the same
ClearCase client. Most of the code paths are the same between the two types of
backends, and the major difference is that the diff output for VersionVault
includes additional information in the headers (the VOB OID for each changed
element, and a JSON-encoded blob to drop into the FileDiff's extra_data in order
to do better post-processing on the diff display).

There's still more to do for unit testing but functionality should be
relatively complete at this point.

Testing Done:
- Posted changes from the current checkout (rbt post with no args) including
  new, modified, deleted, renamed, and binary files.
- Posted changes that spanned multiple VOBs.
- Posted a change that spanned multiple VOBs, including one that was not
  configured as part of the server-side repository. Saw expected
  warning.
- Posted changes for UCM activities for both a single and multiple
  revisions.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/11582/