Don't allow Breezy (Bazaar fork) to match Git repositories.
Breezy, a modern fork of Bazaar, has limited support for working with
Git repositories. If installed, and if SCM scanning is performed, it
will happily state that it can be used with the repository. However,
this doesn't work for our needs, and will cause problems, particularly
if it ends up taking priority over Git.

We now check if it claims the matched repository is Git, and throw out
the result.

Testing Done:
Tested against a local Git repository with SCM scanning. It no longer
came up as a possible match.

Reviewed at https://reviews.reviewboard.org/r/11686/