Add handling for pyenv errors during dependency checks.
When running certain python-based programs, pyenv tries to be "helpful"
if the executable exists in an alternative version of python but not the
current one, and will print an error and give a return code 127. The
previous implementation of our dependency checking only failed if
attempting to run the executable failed in a way that raised an
exception. This change adds special handling to intercept and interpret
the errors that come from pyenv.

Testing Done:
Tested `rbt post` manually in my local integrations repo with the
changes.