Add JSON output to the `rbt list-repo-types` command.
This change adds JSON output for the `rbt list-repo-types` command.
Doing this, I decided that the existing implementation was pretty crummy
in a variety of ways--most specifically, the existence of the
`print_clients` method wasn't helpful, and just meant passing around
output streams too much. I've moved the actual implementation out of
that and into the command, as well as cleaned up things like the text
wrapping. This could probably still be made better with table output or
such but this command is pretty rarely used so it's not worth spending
more time on.

Based on work by Ryan Kang at /r/11573.

Testing Done:
- Ran unit tests.
- Tested `rbt list-repo-types` both with and without `--json`.