Properly handle missing executables when pyenv intercepts.
When pyenv is in use, and the needed executable is only available in a
different Python version, pyenv will intercept the call and print an
error (as well as returning 127). This means that attempts to execute
something like, say, `bzr`, will fail with the following error:

```
pyenv: bzr: command not found

The `bzr' command exists in these Python versions:
  2.7.14
```

This was causing our existing checks to fail with a false positive.

Testing Done:
Tested `rbt post` manually in my local integrations repo with the
changes.

Reviewed at https://reviews.reviewboard.org/r/11196/