Fix the default User-Agent value.
String encoding meant that on Python 3.x, the default user agent was
being sent as "b'RBTools/...'". This change fixes it up to initialize it
as a unicode string, which will get properly normalized on all versions
when we call `str` to insert the header.

Reviewed at https://reviews.reviewboard.org/r/11849/