Improve test skipping for hgsubversion problems.
`hgsubversion` is effectively unmaintained right now. It doesn't work at
all with Python 3, and I'm even struggling to get it running on Python
2.7 with older Mercurial releases.

Assuming it continues in that state, I don't think many people are using
it, and we should probably consider removing it entirely. For now, this
change improves our test skipping machinery to prevent failures when we
can't successfully clone an SVN repository.

Testing Done:
Made a temporary change to get things to proceed halfway in the way that
they are on some setups (past the basic check for `hgsubversion` but
failing when trying to clone). Saw that hgsubversion tests were
successfully skipped rather than causing errors.

Reviewed at https://reviews.reviewboard.org/r/12135/