Enhance the JSON output for rbt status and clean up some non-standard code.
This enhances the results of `rbt status --json` to show additional
information useful for automation.

The results are now in a `review_requests` key, instead of `summaries`,
which better identifies the type of results and allows for enhancement
in the future.

The review request ID (`review_request_id`) and URL
(`review_request_url`) are now provided, instead of just the ID.

The results also now include `has_draft`, `shipit_count`,
`open_issue_count`, `approved`, and `approval_failure` fields, which are
more useful for tools that may want to wrap this command.

It also cleans up some of the code to use more standard terminology
("review request" instead of "request" or "review"), and to fix
documentation and spacing.

Testing Done:
Tested `rbt status --json` and saw the revised fields.

Reviewed at https://reviews.reviewboard.org/r/12177/