Allow the header and footer to be unset in rbt review edit.
`rbt review edit` allowed the header and footer to be set, but they then
couldn't be unset. This was due to checking for truthiness when
determining when to set a new value.

Now it checks for presence in `options` instead. This allows callers to
pass `--header ""` to unset a header.

Testing Done:
Tested setting and then unsetting headers and footers on a review.

Reviewed at https://reviews.reviewboard.org/r/12188/