Add setuptools as an explicit dependency.
We make use of `pkg_resources` for command discovery, which requires the
`setuptools` package, but this isn't guaranteed to be there on all
Python installs. In order to avoid any troubles with installing RBTools,
we now depend explicitly on `setuptools`.

Testing Done:
Tested installing the package. Verified that `setuptools` was in the list
of checked dependencies.

Reviewed at https://reviews.reviewboard.org/r/12192/