Add documentation on rbt post --json and update a JSON key.
This adds a section to the `rbt post` docs that covers the new `--json`
parameter. This documents the success and error schemas and shows off
usage for both.

It also renames the `review_request` key to `review_request_id` in the
JSON payload, in order to be consistent with other commands.

Testing Done:
Built the docs. Checked for build errors and spelling errors.

Reviewed at https://reviews.reviewboard.org/r/12193/