Fix rbt alias --dry-run with system command aliases.
When running `rbt alias --dry-run` on a system command alias (an alias
that executes a command, marked with a leading `!`), the result would be
a list of characters separated by spaces.

This was due to a bad assumption in the code where we assumed the value
we were working with was a list, each representing part of a command
invocation. These sorts of aliases were represented as a string, so we
ended up treating the string as a list of individual characters,
inserting spaces between each.

This change fixes this bug buy only attempting this normalization if
working with a list. Strings are printed as-is.

Testing Done:
Tested with both types of aliases. Saw the expected results.

Tested the links in the new release notes.

Reviewed at https://reviews.reviewboard.org/r/12194/