Update the Windows installer to ship Python 3.8.13.
Previously, the Windows installer was shipping Python 3.8.10, which was
the newest release in the 3.8 series at the time. We're now shipping
3.8.13.

This was chosen over Python 3.9 or 3.10 due to still being compatible
with Windows 7 systems. While we may want to drop Windows 7 soon, we
want to make sure we communicate that correctly before we do it.

There are no official Python 3.8.13 builds, so we've built our own
signed installer using their official process, hosted on S3. This has
been tested and verified to work. When we move to a newer release, we
can go back to consuming official installers.

There are some other fixes in here as well. We no longer block on MD5
verification, which could happen sometimes due to PowerShell looking for
more scripts on stdin. We also have better error handling for the
installation stage, and we output the path to the installer in case we
need to debug things.

Testing Done:
Successfully built the installers. Verified they worked on a Windows
install without Python.

Reviewed at https://reviews.reviewboard.org/r/12203/